+++ composer.json
@@ -15,13 +15,9 @@
     "require": {
         "php": "^8.2",
         "ext-dom": "*",
-        "presttec\/illuminate-contracts": "^10.0|^11.0",
-        "presttec\/illuminate-database": "^10.0|^11.0",
-        "presttec\/illuminate-http": "^10.0|^11.0",
-        "presttec\/illuminate-support": "^10.0|^11.0",
-        "presttec\/illuminate-testing": "^10.0|^11.0",
+        "presttec\/laravel-framework": "^10.44|^11.0",
         "presttec\/mockery": "^1.0",
-        "presttec\/phpunit": "^10.4|^11.0",
+        "presttec\/phpunit": "^10.4|^11.0.1",
         "presttec\/console": "^6.2|^7.0",
         "presttec\/css-selector": "^6.2|^7.0",
         "presttec\/dom-crawler": "^6.2|^7.0",
@@ -29,7 +25,7 @@
         "presttec\/http-kernel": "^6.2|^7.0"
     },
     "require-dev": {
-        "presttec\/laravel-framework": "^10.0|^11.0"
+        "presttec\/testbench-core": "^8.21|^9.0"
     },
     "autoload": {
         "psr-4": {
