+++ phpunit.xml.dist
@@ -10,7 +10,7 @@
 >
     <testsuites>
         <testsuite name="Unit">
-        <directory suffix="Test.php">./tests</directory>
+            <directory suffix="Test.php">./tests</directory>
         </testsuite>
     </testsuites>
 </phpunit>
