+++ src/Constraints/HasSource.php
@@ -2,16 +2,46 @@
 
 namespace Laravel\BrowserKitTesting\Constraints;
 
-use PHPUnit\Runner\Version;
+class HasSource extends PageConstraint
+{
+    /**
+     * The expected HTML source.
+     *
+     * @var string
+     */
+    protected readonly string $source;
 
-if (str_starts_with(Version::series(), '10')) {
-    class HasSource extends PageConstraint
+    /**
+     * Create a new constraint instance.
+     *
+     * @param  string  $source
+     * @return void
+     */
+    public function __construct($source)
     {
-        use Concerns\HasSource;
+        $this->source = $source;
     }
-} else {
-    readonly class HasSource extends PageConstraint
+
+    /**
+     * Check if the source is found in the given crawler.
+     *
+     * @param  \Symfony\Component\DomCrawler\Crawler|string  $crawler
+     * @return bool
+     */
+    protected function matches($crawler): bool
     {
-        use Concerns\HasSource;
+        $pattern = $this->getEscapedPattern($this->source);
+
+        return preg_match("/{$pattern}/i", $this->html($crawler));
+    }
+
+    /**
+     * Returns a string representation of the object.
+     *
+     * @return string
+     */
+    public function toString(): string
+    {
+        return "the HTML [{$this->source}]";
     }
 }
