+++ src/Constraints/HasValue.php
@@ -2,16 +2,73 @@
 
 namespace Laravel\BrowserKitTesting\Constraints;
 
-use PHPUnit\Runner\Version;
+use Symfony\Component\DomCrawler\Crawler;
 
-if (str_starts_with(Version::series(), '10')) {
-    class HasValue extends FormFieldConstraint
+class HasValue extends FormFieldConstraint
+{
+    /**
+     * Get the valid elements.
+     *
+     * @return string
+     */
+    protected function validElements()
     {
-        use Concerns\HasValue;
+        return 'input,textarea';
     }
-} else {
-    readonly class HasValue extends FormFieldConstraint
+
+    /**
+     * Check if the input contains the expected value.
+     *
+     * @param  \Symfony\Component\DomCrawler\Crawler|string  $crawler
+     * @return bool
+     */
+    public function matches($crawler): bool
     {
-        use Concerns\HasValue;
+        $crawler = $this->crawler($crawler);
+
+        return $this->getInputOrTextAreaValue($crawler) == $this->value;
+    }
+
+    /**
+     * Get the value of an input or textarea.
+     *
+     * @param  \Symfony\Component\DomCrawler\Crawler  $crawler
+     * @return string
+     *
+     * @throws \PHPUnit\Framework\ExpectationFailedException
+     */
+    public function getInputOrTextAreaValue(Crawler $crawler)
+    {
+        $field = $this->field($crawler);
+
+        return $field->nodeName() == 'input'
+            ? $field->attr('value')
+            : $field->text();
+    }
+
+    /**
+     * Return the description of the failure.
+     *
+     * @return string
+     */
+    protected function getFailureDescription()
+    {
+        return sprintf(
+            'the field [%s] contains the expected value [%s]',
+            $this->selector, $this->value
+        );
+    }
+
+    /**
+     * Returns the reversed description of the failure.
+     *
+     * @return string
+     */
+    protected function getReverseFailureDescription()
+    {
+        return sprintf(
+            'the field [%s] does not contain the expected value [%s]',
+            $this->selector, $this->value
+        );
     }
 }
