+++ src/Constraints/IsChecked.php
@@ -2,16 +2,59 @@
 
 namespace Laravel\BrowserKitTesting\Constraints;
 
-use PHPUnit\Runner\Version;
-
-if (str_starts_with(Version::series(), '10')) {
-    class IsChecked extends FormFieldConstraint
+class IsChecked extends FormFieldConstraint
+{
+    /**
+     * Create a new constraint instance.
+     *
+     * @param  string  $selector
+     * @return void
+     */
+    public function __construct($selector)
     {
-        use Concerns\IsChecked;
+        parent::__construct($selector, null);
     }
-} else {
-    readonly class IsChecked extends FormFieldConstraint
+
+    /**
+     * Get the valid elements.
+     *
+     * @return string
+     */
+    protected function validElements()
     {
-        use Concerns\IsChecked;
+        return "input[type='checkbox']";
+    }
+
+    /**
+     * Determine if the checkbox is checked.
+     *
+     * @param  \Symfony\Component\DomCrawler\Crawler|string  $crawler
+     * @return bool
+     */
+    public function matches($crawler): bool
+    {
+        $crawler = $this->crawler($crawler);
+
+        return ! is_null($this->field($crawler)->attr('checked'));
+    }
+
+    /**
+     * Return the description of the failure.
+     *
+     * @return string
+     */
+    protected function getFailureDescription()
+    {
+        return "the checkbox [{$this->selector}] is checked";
+    }
+
+    /**
+     * Returns the reversed description of the failure.
+     *
+     * @return string
+     */
+    protected function getReverseFailureDescription()
+    {
+        return "the checkbox [{$this->selector}] is not checked";
     }
 }
